LayoutTests/imported/w3c/web-platform-tests/webrtc-encoded-transform/script-transform-generateKeyFrame.https-expected.txt



PASS generateKeyFrame() throws for audio
PASS generateKeyFrame(null) resolves for video sender, and throws for video receiver
FAIL generateKeyFrame throws NotAllowedError for invalid rid assert_equals: expected "failure" but got "success"
FAIL generateKeyFrame throws NotFoundError for unknown rid assert_equals: expected "failure" but got "success"
PASS generateKeyFrame throws for unset transforms
PASS generateKeyFrame timestamp should advance
PASS await generateKeyFrame, await generateKeyFrame should see an increase in count of keyframes
FAIL generateKeyFrame rejects when the sender is negotiated inactive, and resumes succeeding when negotiated back to active assert_equals: Message: Timed out after waiting for 8000 ms expected "InvalidStateError" but got "TimeoutError"
PASS generateKeyFrame rejects when the sender is stopped, even without negotiation
FAIL generateKeyFrame rejects with a null track assert_equals: Message: Timed out after waiting for 8000 ms expected "InvalidStateError" but got "TimeoutError"
